refactor(#97): remove server pkg, fix problem with config load, update bash script readability.
