+++ app-config.example.yaml
@@ -6,9 +6,3 @@ server:
     app_name: Overlay API v1.0.0
     port: 3000
     server_header: Overlay API
-mongo:
-    auth_db: admin
-    database: overlay
-    password: ""
-    uri: mongodb://localhost:27017
-    username: ""
