+++ cmd/srv/main.go
@@ -3,16 +3,15 @@ package main
 import (
 	"context"
 	"flag"
-	"net/http"
 
-	"github.com/4chain-ag/go-overlay-services/pkg/server"
-	"github.com/4chain-ag/go-overlay-services/pkg/server/config"
-	"github.com/4chain-ag/go-overlay-services/pkg/server/config/loaders"
+	"github.com/4chain-ag/go-overlay-services/pkg/server2"
+	"github.com/4chain-ag/go-overlay-services/pkg/server2/config"
+	"github.com/4chain-ag/go-overlay-services/pkg/server2/config/loaders"
 	"github.com/gookit/slog"
 )
 
 func main() {
-	configPath := flag.String("c", loaders.DefaultConfigFilePath, "Path to the configuration file")
+	configPath := flag.String("config", loaders.DefaultConfigFilePath, "Path to the configuration file")
 	flag.Parse()
 
 	loader := loaders.NewLoader(config.NewDefault, "OVERLAY")
@@ -29,37 +28,6 @@ func main() {
 		slog.Fatalf("failed to pretty print config: %v", err)
 	}
 
-	opts := []server.HTTPOption{
-		server.WithMiddleware(loggingMiddleware),
-		server.WithConfig(&cfg.Server),
-	}
-
-	httpAPI, err := server.New(opts...)
-	if err != nil {
-		slog.Fatalf("Failed to create HTTP server: %v", err)
-	}
-
-	// Graceful shutdown handling
-	ctx := context.Background()
-	idleConnsClosed := httpAPI.StartWithGracefulShutdown(ctx)
-	<-idleConnsClosed
-	slog.Info("Server shutdown completed.")
-}
-
-// loggingMiddleware is a custom definition of the logging middleware format accepted by the HTTP API.
-func loggingMiddleware(next http.Handler) http.Handler {
-	slog.SetLogLevel(slog.DebugLevel)
-	slog.SetFormatter(slog.NewJSONFormatter(func(f *slog.JSONFormatter) {
-		f.PrettyPrint = true
-	}))
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		logger := slog.WithFields(slog.M{
-			"category":    "service",
-			"method":      r.Method,
-			"remote-addr": r.RemoteAddr,
-			"request-uri": r.RequestURI,
-		})
-		logger.Info("log-line")
-		next.ServeHTTP(w, r)
-	})
+	srv := server2.New(server2.WithConfig(&cfg.Server))
+	<-srv.ListenAndServe(context.Background())
 }
