+++ docker-compose.yaml
@@ -6,6 +6,6 @@ services:
     ports:
       - 8080:8080
     environment:
-      SWAGGER_JSON: /foo/bundled.yaml
+      SWAGGER_JSON: /foo/overlay-services.yaml
     volumes:
       - ./docs/api/openapi:/foo:ro
